+++ lib/td_df/templates/template.ex
@@ -7,6 +7,31 @@ defmodule TdDf.Templates.Template do
 
   alias TdDf.Templates
 
+  @validations [
+    :repeated_group_names,
+    :repeated_names,
+    :name_and_types,
+    :subscribable,
+    :group_names,
+    :field_names,
+    :field_label,
+    :widget,
+    :type,
+    :cardinality,
+    :user_roles,
+    :group_roles,
+    :type_hierarchy,
+    :type_table,
+    :dropdown_string_value,
+    :depending_domain,
+    :depending_domain_list,
+    :depending_field,
+    :fixed_list,
+    :key_value_list,
+    :conditional_visibility,
+    :mandatory_depending
+  ]
+
   schema "templates" do
     field(:content, {:array, :map})
     field(:label, :string)
@@ -30,43 +55,53 @@ defmodule TdDf.Templates.Template do
          %{} = params,
          attrs \\ [:label, :name, :content, :scope, :subscope]
        ) do
-    struct
-    |> cast(params, attrs)
-    |> validate_required([:label, :name, :content, :scope])
-    |> validate_repeated_group_names()
-    |> validate_repeated_names()
-    |> validate_name_and_types(struct)
-    |> validate_subscribable()
-    |> unique_constraint(:name)
+    changeset =
+      struct
+      |> cast(params, attrs)
+      |> validate_required([:label, :name, :content, :scope])
+      |> unique_constraint(:name)
+
+    Enum.reduce(@validations, changeset, fn validation, changeset ->
+      validate(changeset, validation, struct)
+    end)
   end
 
-  defp validate_repeated_group_names(%{valid?: true} = changeset) do
+  defp validate(%{valid?: true} = changeset, :repeated_group_names, _) do
     changeset
     |> get_field(:content)
     |> Enum.frequencies_by(&Map.get(&1, "name"))
     |> Enum.max_by(fn {_, count} -> count end, fn -> {:none, 0} end)
     |> case do
+      {:none, 0} -> add_error(changeset, :content, "invalid_content.no_groups")
       {name, count} when count > 1 -> add_error(changeset, :content, "repeated.group", name: name)
       _ -> changeset
     end
   end
 
-  defp validate_repeated_group_names(changeset), do: changeset
+  defp validate(%{valid?: true} = changeset, :repeated_names, _) do
+    content = get_field(changeset, :content)
 
-  defp validate_repeated_names(%{valid?: true} = changeset) do
-    changeset
-    |> flatten_content_fields()
-    |> Enum.frequencies_by(&Map.get(&1, "name"))
-    |> Enum.max_by(fn {_, count} -> count end, fn -> {:none, 0} end)
-    |> case do
-      {name, count} when count > 1 -> add_error(changeset, :content, "repeated.field", name: name)
-      _ -> changeset
+    if Enum.any?(content, &match?(%{"fields" => nil}, &1)) do
+      add_error(changeset, :content, "invalid_group.no_fields")
+    else
+      changeset
+      |> flatten_content_fields()
+      |> Enum.frequencies_by(&Map.get(&1, "name"))
+      |> Enum.max_by(fn {_, count} -> count end, fn -> {:none, 0} end)
+      |> case do
+        {:none, 0} ->
+          add_error(changeset, :content, "invalid_group.no_fields")
+
+        {name, count} when count > 1 ->
+          add_error(changeset, :content, "repeated.field", name: name)
+
+        _ ->
+          changeset
+      end
     end
   end
 
-  defp validate_repeated_names(changeset), do: changeset
-
-  defp validate_name_and_types(%{valid?: true} = changeset, %{id: id}) do
+  defp validate(%{valid?: true} = changeset, :name_and_types, %{id: id}) do
     templates =
       Enum.filter(Templates.list_templates(), fn template -> Map.get(template, :id) != id end)
 
@@ -86,9 +121,7 @@ defmodule TdDf.Templates.Template do
     end
   end
 
-  defp validate_name_and_types(changeset, _), do: changeset
-
-  defp validate_subscribable(%{valid?: true} = changeset) do
+  defp validate(%{valid?: true} = changeset, :subscribable, _) do
     changeset
     |> flatten_content_fields()
     |> Enum.filter(&Map.get(&1, "subscribable"))
@@ -102,7 +135,299 @@ defmodule TdDf.Templates.Template do
     end
   end
 
-  defp validate_subscribable(changeset), do: changeset
+  defp validate(%{valid?: true} = changeset, :group_names, _) do
+    changeset
+    |> get_field(:content)
+    |> Enum.filter(&empty_or_nil?(Map.get(&1, "name")))
+    |> case do
+      [] ->
+        changeset
+
+      [field | _] ->
+        add_error(changeset, :content, "invalid.group.name", name: Map.get(field, "name"))
+    end
+  end
+
+  defp validate(%{valid?: true} = changeset, :field_names, _) do
+    changeset
+    |> flatten_content_fields()
+    |> Enum.filter(&empty_or_nil?(Map.get(&1, "name")))
+    |> case do
+      [] ->
+        changeset
+
+      [field | _] ->
+        add_error(changeset, :content, "invalid.field.name", name: Map.get(field, "name"))
+    end
+  end
+
+  defp validate(%{valid?: true} = changeset, :field_label, _) do
+    changeset
+    |> flatten_content_fields()
+    |> Enum.filter(&empty_or_nil?(Map.get(&1, "label")))
+    |> case do
+      [] ->
+        changeset
+
+      [field | _] ->
+        add_error(changeset, :content, "invalid.field.label", name: Map.get(field, "name"))
+    end
+  end
+
+  defp validate(%{valid?: true} = changeset, :widget, _) do
+    changeset
+    |> flatten_content_fields()
+    |> Enum.filter(&empty_or_nil?(Map.get(&1, "widget")))
+    |> case do
+      [] ->
+        changeset
+
+      [field | _] ->
+        add_error(changeset, :content, "invalid.field.widget", name: Map.get(field, "name"))
+    end
+  end
+
+  defp validate(%{valid?: true} = changeset, :type, _) do
+    changeset
+    |> flatten_content_fields()
+    |> Enum.filter(&empty_or_nil?(Map.get(&1, "type")))
+    |> case do
+      [] ->
+        changeset
+
+      [field | _] ->
+        add_error(changeset, :content, "invalid.field.type", name: Map.get(field, "name"))
+    end
+  end
+
+  defp validate(%{valid?: true} = changeset, :cardinality, _) do
+    changeset
+    |> flatten_content_fields()
+    |> Enum.filter(&empty_or_nil?(Map.get(&1, "cardinality")))
+    |> case do
+      [] ->
+        changeset
+
+      [field | _] ->
+        add_error(changeset, :content, "invalid.field.cardinality", name: Map.get(field, "name"))
+    end
+  end
+
+  defp validate(%{valid?: true} = changeset, :user_roles, _) do
+    changeset
+    |> flatten_content_fields()
+    |> Enum.filter(
+      &(Map.get(&1, "type") === "user" &&
+          empty_or_nil?(Map.get(&1, "values")["role_users"]))
+    )
+    |> case do
+      [] ->
+        changeset
+
+      [field | _] ->
+        add_error(changeset, :content, "invalid.field.values.role_users",
+          name: Map.get(field, "name")
+        )
+    end
+  end
+
+  defp validate(%{valid?: true} = changeset, :group_roles, _) do
+    changeset
+    |> flatten_content_fields()
+    |> Enum.filter(
+      &(Map.get(&1, "type") === "user_group" &&
+          empty_or_nil?(Map.get(&1, "values")["role_groups"]))
+    )
+    |> case do
+      [] ->
+        changeset
+
+      [field | _] ->
+        add_error(changeset, :content, "invalid.field.values.role_groups",
+          name: Map.get(field, "name")
+        )
+    end
+  end
+
+  defp validate(%{valid?: true} = changeset, :type_hierarchy, _) do
+    changeset
+    |> flatten_content_fields()
+    |> Enum.filter(
+      &(Map.get(&1, "type") === "hierarchy" && empty_or_nil?(Map.get(&1, "values")["hierarchy"]))
+    )
+    |> case do
+      [] ->
+        changeset
+
+      [field | _] ->
+        add_error(changeset, :content, "invalid.field.values.hierarchy",
+          name: Map.get(field, "name")
+        )
+    end
+  end
+
+  defp validate(%{valid?: true} = changeset, :type_table, _) do
+    changeset
+    |> flatten_content_fields()
+    |> Enum.filter(
+      &(Map.get(&1, "type") === "table" && empty_or_nil?(Map.get(&1, "values")["table_columns"]))
+    )
+    |> case do
+      [] ->
+        changeset
+
+      [field | _] ->
+        add_error(changeset, :content, "invalid.field.values.table_columns",
+          name: Map.get(field, "name")
+        )
+    end
+  end
+
+  defp validate(%{valid?: true} = changeset, :dropdown_string_value, _) do
+    changeset
+    |> flatten_content_fields()
+    |> Enum.filter(
+      &(Map.get(&1, "widget") in ["dropdown", "radio", "checkout"] &&
+          Map.get(&1, "type") === "string" && empty_or_nil?(Map.get(&1, "values")))
+    )
+    |> case do
+      [] ->
+        changeset
+
+      [field | _] ->
+        add_error(changeset, :content, "invalid.field.values.type", name: Map.get(field, "name"))
+    end
+  end
+
+  defp validate(%{valid?: true} = changeset, :depending_domain, _) do
+    changeset
+    |> flatten_content_fields()
+    |> Enum.filter(
+      &(Map.get(&1, "widget") in ["dropdown", "radio", "checkout"] &&
+          Map.get(&1, "type") === "string" && Map.get(&1, "values")["domain"] === %{})
+    )
+    |> case do
+      [] ->
+        changeset
+
+      [field | _] ->
+        add_error(changeset, :content, "invalid.field.values.domain.field",
+          name: Map.get(field, "name")
+        )
+    end
+  end
+
+  defp validate(%{valid?: true} = changeset, :depending_domain_list, _) do
+    changeset
+    |> flatten_content_fields()
+    |> Enum.filter(fn field ->
+      Map.get(field, "widget") in ["dropdown", "radio", "checkout"] &&
+        Map.get(field, "type") === "string" &&
+        Map.get(field, "values")["domain"] &&
+        field["values"]["domain"]
+        |> Map.values()
+        |> Enum.any?(&empty_or_nil?(&1))
+    end)
+    |> case do
+      [] ->
+        changeset
+
+      [field | _] ->
+        add_error(changeset, :content, "invalid.field.values.domain.list",
+          name: Map.get(field, "name")
+        )
+    end
+  end
+
+  defp validate(%{valid?: true} = changeset, :depending_field, _) do
+    changeset
+    |> flatten_content_fields()
+    |> Enum.filter(fn field ->
+      Map.get(field, "widget") in ["dropdown", "radio", "checkout"] &&
+        Map.get(field, "type") === "string" &&
+        Map.get(field, "values")["switch"] &&
+        field["values"]["switch"]
+        |> Map.values()
+        |> Enum.any?(&empty_or_nil?(&1))
+    end)
+    |> case do
+      [] ->
+        changeset
+
+      [field | _] ->
+        add_error(changeset, :content, "invalid.field.values.switch",
+          name: Map.get(field, "name")
+        )
+    end
+  end
+
+  defp validate(%{valid?: true} = changeset, :fixed_list, _) do
+    changeset
+    |> flatten_content_fields()
+    |> Enum.filter(
+      &(Map.get(&1, "widget") in ["dropdown", "radio", "checkout"] &&
+          Map.get(&1, "type") === "string" && Map.get(&1, "values")["fixed"] === [])
+    )
+    |> case do
+      [] ->
+        changeset
+
+      [field | _] ->
+        add_error(changeset, :content, "invalid.field.values.fixed", name: Map.get(field, "name"))
+    end
+  end
+
+  defp validate(%{valid?: true} = changeset, :key_value_list, _) do
+    changeset
+    |> flatten_content_fields()
+    |> Enum.filter(
+      &(Map.get(&1, "widget") in ["dropdown", "radio", "checkout"] &&
+          Map.get(&1, "type") === "string" && Map.get(&1, "values")["fixed_tuple"] === [])
+    )
+    |> case do
+      [] ->
+        changeset
+
+      [field | _] ->
+        add_error(changeset, :content, "invalid.field.values.fixed_tuple",
+          name: Map.get(field, "name")
+        )
+    end
+  end
+
+  defp validate(%{valid?: true} = changeset, :conditional_visibility, _) do
+    changeset
+    |> flatten_content_fields()
+    |> Enum.filter(&(Map.get(&1, "depends") && empty_or_nil?(Map.get(&1, "depends")["to_be"])))
+    |> case do
+      [] ->
+        changeset
+
+      [field | _] ->
+        add_error(changeset, :content, "invalid.field.depends.to_be",
+          name: Map.get(field, "name")
+        )
+    end
+  end
+
+  defp validate(%{valid?: true} = changeset, :mandatory_depending, _) do
+    changeset
+    |> flatten_content_fields()
+    |> Enum.filter(
+      &(Map.get(&1, "mandatory") && empty_or_nil?(Map.get(&1, "mandatory")["to_be"]))
+    )
+    |> case do
+      [] ->
+        changeset
+
+      [field | _] ->
+        add_error(changeset, :content, "invalid.field.mandatory.to_be",
+          name: Map.get(field, "name")
+        )
+    end
+  end
+
+  defp validate(changeset, _, _), do: changeset
 
   defp flatten_content_fields(changeset) do
     changeset
@@ -140,4 +465,8 @@ defmodule TdDf.Templates.Template do
   defp fixed_values(%{"values" => %{"fixed" => _}}), do: true
   defp fixed_values(%{"values" => %{"fixed_tuple" => _}}), do: true
   defp fixed_values(_), do: false
+
+  defp empty_or_nil?(value) do
+    value in [nil, "", [], %{}]
+  end
 end